Fix potential runtime panic caused by closing HTTP response body on error
	If the returned error is nil, the Response will contain
	a non-nil Body which the user is expected to close.

Based on the above (https://golang.org/pkg/net/http/#Client.Do),
we do not need to check if body != nil before closing.

Furthermore, we should only ever defer a close after checking
for an error. If there are any errors, we do not need to close
as the body is nil.