Fix runtime panic error when trying to change the scheme of an invalid URL
`17fa918` added the ability to proxy protocol-relative URIs by
setting the `Scheme` to `https` by default if it is not set.
Unfortunately, it did so prior to checking if there were
any errors parsing the URL. Consequently, the following error
would be returned if an invalid URL is parsed:

	runtime error: invalid memory address or nil pointer dereference

This commit fixes it by setting the `Scheme` after checking
for errors.

---

Closes CAMO-3