Do not override the original `io.ReadCloser` with the gzip reader
The parent should handle the 'closing' of the original / given
`io.ReadCloser`. But, as we are 'creating' a new reader for
gzip-encoded content, the function doing that should be
responsible for closing the reader after all bytes are read.
To be explicit, we should set the new reader to a different
variable, and close it accordingly. Also, prior to this commit,
we did not do any checking to see if there were errors on closing.
This commit does that.